Synchronize Python 3 to Makefile (#87)
* Build and install Python 3 module with CMake
Use links, make_pycbf.py, and SWIG to generate interfaces from
CBFlib.html.  Compile with SWIG_PYTHON_STRICT_BYTE_CHAR, which causes it
to be multiply defined.

* Make Python 3 tests take paths of in- and output files from command line
This reduces the constraints on the buildsystem and makes it easier to
clean up after tests: tests do not need to run in a subdirectory of the
input data directory.

* Avoid inconsequential output in Python 3 tests
Do not output lines that would have be removed before subsequent
comparisons.

* Add Python 3 tests in CMake
Several tests rely on cbf_standardize_numbers, which in turn needs
fgetln(3).

* Install python3-numpy-dev dependency for GitHub action
While here, also set the maximum number of concurrent CMake processes to
`nproc`.